* Better handling of duplicates     * Better handling of icons     * Drag and drop icon support     * Selects icons from system icon theme     * Removed default quickly template menus     * Added some stub menu entries     * Enabled some of webkit features  like webgl, html5 db etc     * Now each app has a different session

+++ fogger_lib/FogApp.py
@@ -1,17 +1,22 @@
 import urllib2
 import urlparse
 import os
+import logging
 import simplejson as json
+from hashlib import md5
 
-from gi.repository import GLib
+from gi.repository import GLib, WebKit, Soup
 
 from fogger import AppWindow
-from fogger_lib.helpers import get_or_create_directory, get_media_file
+from fogger_lib.helpers import get_or_create_directory
 from . foggerconfig import get_data_file
+
+logger = logging.getLogger('fogger_lib')
 
 
 __all__ = ('FogApp', 'FogAppManager', 'app_manager')
 op = os.path
+
 APP_PATH = op.join(GLib.get_user_data_dir(), 'fogapps')
 CONF_PATH = op.join(GLib.get_user_config_dir(), 'fogger')
 
@@ -21,6 +26,8 @@ class FogApp:
     url = None
     window_size = (800, 600,)
     maximized = False
+    icon = None
+    uuid = None
     path = ''
 
     def __init__(self, path=None):
@@ -31,8 +38,22 @@ class FogApp:
             state = json.loads(open(op.join(self.path, 'app.json'), 'r').read())
             self.name = state['name']
             self.url = state['url']
+            self.uuid = state['uuid']
+            self.icon = state['icon']
             self.window_size = state['window_size']
             self.maximized = state['maximized']
+
+    def get_stylesheet(self):
+        return op.join(self.path, 'styles', 'style.css')
+
+    def get_desktop_file(self):
+        return op.join(GLib.get_user_data_dir(), 'applications', '%s.desktop' % self.uuid)
+
+    def setup_webkit_session(self):
+        session = WebKit.get_default_session()
+        cache = get_or_create_directory(os.path.join(GLib.get_user_cache_dir(), 'fogger', self.uuid))
+        cookie_jar = Soup.CookieJarText.new(op.join(cache, 'WebkitSession'), False)
+        session.add_feature(cookie_jar)
 
     def set_url(self, url):
         self.url = url
@@ -48,6 +69,8 @@ class FogApp:
         state = {
             'name': self.name,
             'url': self.url,
+            'uuid': self.uuid,
+            'icon': self.icon,
             'window_size': self.window_size,
             'maximized': self.maximized,
         }
@@ -57,6 +80,7 @@ class FogApp:
         handle.close()
 
     def run(self):
+        self.setup_webkit_session()
         self.window = AppWindow.AppWindow()
         self.window.run_app(self)
 
@@ -69,45 +93,53 @@ class FogAppManager:
         if op.exists(path):
             self.apps = json.loads(open(path, 'r').read())
 
-    def get(self, name):
-        path = self.apps.get(name)
-        if path:
+    def get(self, uuid):
+        #path = self.apps.get(uuid)
+        path = op.join(APP_PATH, uuid)
+        if op.exists(path):
             return FogApp(path)
         else:
+            logger.error('No such app: %s' % uuid)
             return None
 
     def create(self, name, url, icon):
-        path = self._setup_app_dir(name)
+        print icon, '<<<'
+        uuid = md5(name).hexdigest()
+        path = self._setup_app_dir(uuid)
         app = FogApp()
         app.name = name
         app.path = path
+        app.uuid = uuid
+        app.icon = icon
         app.set_url(url)
         app.save()
-        icon = self._setup_icon(name, icon)
-        self._create_desktop_file(name, icon)
-        self.apps[name] = path
-        self.save()
+        icon = self._setup_icon(uuid, icon)
+        self._create_desktop_file(name, uuid, icon)
+        #self.apps[uuid] = path
+        #self.save()
         return app
 
-    def _setup_icon(self, name, icon):
-        _, ext = op.splitext(icon)
-        path = op.join(APP_PATH, name, 'icon%s' % ext)
-        data = open(icon).read()
-        open(path, 'w').write(data)
-        return path
+    def _setup_icon(self, uuid, icon):
+        if op.exists(icon):
+            _, ext = op.splitext(icon)
+            path = op.join(APP_PATH, uuid, 'icon%s' % ext)
+            data = open(icon).read()
+            open(path, 'w').write(data)
+            return path
+        else:
+            return icon
 
-    def _create_desktop_file(self, name, icon):
-        print get_data_file('templates/fogapp.desktop.in'), '<<<<<<'
-        desktop_tmpl = open(get_data_file('templates/fogapp.desktop.in')).read()
-        desktop_tmpl = desktop_tmpl % {'name': name, 'icon': icon}
-        desktop_file_path = op.join(GLib.get_user_data_dir(), 'applications', '%s.desktop' % name)
+    def _create_desktop_file(self, name, uuid, icon):
+        desktop_tmpl = open(get_data_file('templates/fogapp.desktop.tmpl')).read()
+        desktop_tmpl = desktop_tmpl % {'name': name, 'icon': icon, 'uuid': uuid}
+        desktop_file_path = op.join(GLib.get_user_data_dir(), 'applications', '%s.desktop' % uuid)
         desktop_file = open(desktop_file_path, 'w')
         desktop_file.write(desktop_tmpl)
         desktop_file.close()
         os.chmod(desktop_file_path, 0755)
 
-    def _setup_app_dir(self, name):
-        path = get_or_create_directory(op.join(APP_PATH, name))
+    def _setup_app_dir(self, uuid):
+        path = get_or_create_directory(op.join(APP_PATH, uuid))
         get_or_create_directory(op.join(path, 'scripts'))
         get_or_create_directory(op.join(path, 'styles'))
         return path
